refactor(correct-score): correct naming of pure version
