+++ components/common/modal/CorrectScore.stories.js
@@ -1,15 +1,15 @@
-import PureCorrectScore from './PureCorrectScore.vue'
+import CorrectScorePure from './CorrectScorePure.vue'
 
 export default {
-  component: PureCorrectScore,
-  title: 'PureCorrectScore',
+  component: CorrectScorePure,
+  title: 'CorrectScorePure',
   exludeStories: /.*Data$/,
 }
 
 const Template = (args, { argTypes }) => ({
-  components: { PureCorrectScore },
+  components: { CorrectScorePure },
   props: Object.keys(argTypes),
-  template: '<pureCorrectScore v-bind="$props"/>',
+  template: '<correctScorePure v-bind="$props"/>',
 })
 export const Default = Template.bind({})
 Default.args = {
